python_service/requirements.txt
torch~=2.5.1+cu121
transformers~=4.50.3
pillow~=11.1.0
opencv-python~=4.11.0.86
ultralytics~=8.3.105
flask~=3.1.0
torchvision~=0.20.1+cu121